# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "RSVP System",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/events": {
   "get": {
    "summary": "List Events",
    "operationId": "list_events_api_events_get",
    "parameters": [
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/EventOut"
         },
         "title": "Response List Events Api Events Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "summary": "Create Event",
    "operationId": "create_event_api_events_post",
    "parameters": [
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EventIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/events/{event_id}": {
   "get": {
    "summary": "Read Event",
    "operationId": "read_event_api_events__event_id__get",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Event Id"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "summary": "Update Event",
    "operationId": "update_event_api_events__event_id__put",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Event Id"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/EventIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/EventOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "summary": "Delete Event",
    "operationId": "delete_event_api_events__event_id__delete",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Event Id"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/events/{event_id}/imports/parse": {
   "post": {
    "summary": "Parse",
    "operationId": "parse_api_events__event_id__imports_parse_post",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Event Id"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_parse_api_events__event_id__imports_parse_post"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ParsedSheet"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/events/{event_id}/imports/preview": {
   "post": {
    "summary": "Preview",
    "operationId": "preview_api_events__event_id__imports_preview_post",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Event Id"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PreviewRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Preview"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/events/{event_id}/imports/commit": {
   "post": {
    "summary": "Commit",
    "operationId": "commit_api_events__event_id__imports_commit_post",
    "parameters": [
     {
      "name": "event_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Event Id"
      }
     },
     {
      "name": "session",
      "in": "cookie",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Session"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CommitRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CommitResult"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "Body_parse_api_events__event_id__imports_parse_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_parse_api_events__event_id__imports_parse_post"
   },
   "CommitRequest": {
    "properties": {
     "mapping": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Mapping"
     },
     "rows": {
      "items": {
       "items": {
        "type": "string"
       },
       "type": "array"
      },
      "type": "array",
      "title": "Rows"
     },
     "actions": {
      "additionalProperties": {
       "type": "string",
       "enum": [
        "skip",
        "update",
        "new"
       ]
      },
      "type": "object",
      "title": "Actions",
      "default": {}
     }
    },
    "type": "object",
    "required": [
     "mapping",
     "rows"
    ],
    "title": "CommitRequest"
   },
   "CommitResult": {
    "properties": {
     "created": {
      "type": "integer",
      "title": "Created"
     },
     "updated": {
      "type": "integer",
      "title": "Updated"
     },
     "skipped": {
      "type": "integer",
      "title": "Skipped"
     }
    },
    "type": "object",
    "required": [
     "created",
     "updated",
     "skipped"
    ],
    "title": "CommitResult"
   },
   "Conflict": {
    "properties": {
     "invitation_id": {
      "type": "integer",
      "title": "Invitation Id"
     },
     "invitation_name": {
      "type": "string",
      "title": "Invitation Name"
     }
    },
    "type": "object",
    "required": [
     "invitation_id",
     "invitation_name"
    ],
    "title": "Conflict"
   },
   "ContactIn": {
    "properties": {
     "phone": {
      "type": "string",
      "maxLength": 50,
      "minLength": 1,
      "title": "Phone"
     },
     "member_index": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Member Index"
     }
    },
    "type": "object",
    "required": [
     "phone"
    ],
    "title": "ContactIn"
   },
...
   "EventIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Name"
     },
     "type": {
      "type": "string",
      "maxLength": 50,
      "title": "Type",
      "default": "wedding"
     },
     "date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date"
     },
     "time": {
      "anyOf": [
       {
        "type": "string",
        "format": "time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Time"
     },
     "venue": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Venue"
     },
     "address": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 300
       },
       {
        "type": "null"
       }
      ],
      "title": "Address"
     },
     "waze_url": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 500
       },
       {
        "type": "null"
       }
      ],
      "title": "Waze Url"
     },
     "maps_url": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 500
       },
       {
        "type": "null"
       }
      ],
      "title": "Maps Url"
     },
     "hosts": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 300
       },
       {
        "type": "null"
       }
      ],
      "title": "Hosts"
     },
     "rsvp_deadline": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rsvp Deadline"
     },
     "default_language": {
      "type": "string",
      "enum": [
       "he",
       "en"
      ],
      "title": "Default Language",
      "default": "he"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "EventIn"
   },
   "EventOut": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Name"
     },
     "type": {
      "type": "string",
      "maxLength": 50,
      "title": "Type",
      "default": "wedding"
     },
     "date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Date"
     },
     "time": {
      "anyOf": [
       {
        "type": "string",
        "format": "time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Time"
     },
     "venue": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Venue"
     },
     "address": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 300
       },
       {
        "type": "null"
       }
      ],
      "title": "Address"
     },
     "waze_url": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 500
       },
       {
        "type": "null"
       }
      ],
      "title": "Waze Url"
     },
     "maps_url": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 500
       },
       {
        "type": "null"
       }
      ],
      "title": "Maps Url"
     },
     "hosts": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 300
       },
       {
        "type": "null"
       }
      ],
      "title": "Hosts"
     },
     "rsvp_deadline": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rsvp Deadline"
     },
     "default_language": {
      "type": "string",
      "enum": [
       "he",
       "en"
      ],
      "title": "Default Language",
      "default": "he"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "id": {
      "type": "integer",
      "title": "Id"
     }
    },
    "type": "object",
    "required": [
     "name",
     "id"
    ],
    "title": "EventOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "InvitationIn": {
    "properties": {
     "display_name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Display Name"
     },
     "group_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 100
       },
       {
        "type": "null"
       }
      ],
      "title": "Group Name"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "members": {
      "items": {
       "$ref": "#/components/schemas/MemberIn"
      },
      "type": "array",
      "title": "Members",
      "default": []
     },
     "contacts": {
      "items": {
       "$ref": "#/components/schemas/ContactIn"
      },
      "type": "array",
      "title": "Contacts",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "display_name"
    ],
    "title": "InvitationIn"
   },
...
   "Issue": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "level": {
      "type": "string",
      "enum": [
       "error",
       "warning",
       "conflict"
      ],
      "title": "Level"
     },
     "detail": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Detail"
     }
    },
    "type": "object",
    "required": [
     "code",
     "level"
    ],
    "title": "Issue"
   },
   "MemberIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 200,
      "minLength": 1,
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "MemberIn"
   },
...
   "ParsedSheet": {
    "properties": {
     "headers": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Headers"
     },
     "rows": {
      "items": {
       "items": {
        "type": "string"
       },
       "type": "array"
      },
      "type": "array",
      "title": "Rows"
     },
     "mapping": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Mapping"
     }
    },
    "type": "object",
    "required": [
     "headers",
     "rows",
     "mapping"
    ],
    "title": "ParsedSheet"
   },
   "Preview": {
    "properties": {
     "rows": {
      "items": {
       "$ref": "#/components/schemas/PreviewRow"
      },
      "type": "array",
      "title": "Rows"
     },
     "counts": {
      "additionalProperties": {
       "type": "integer"
      },
      "type": "object",
      "title": "Counts"
     }
    },
    "type": "object",
    "required": [
     "rows",
     "counts"
    ],
    "title": "Preview"
   },
   "PreviewRequest": {
    "properties": {
     "mapping": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Mapping"
     },
     "rows": {
      "items": {
       "items": {
        "type": "string"
       },
       "type": "array"
      },
      "type": "array",
      "title": "Rows"
     }
    },
    "type": "object",
    "required": [
     "mapping",
     "rows"
    ],
    "title": "PreviewRequest"
   },
   "PreviewRow": {
    "properties": {
     "index": {
      "type": "integer",
      "title": "Index"
     },
     "draft": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/InvitationIn"
       },
       {
        "type": "null"
       }
      ]
     },
     "status": {
      "type": "string",
      "enum": [
       "ok",
       "warning",
       "conflict",
       "error"
      ],
      "title": "Status"
     },
     "issues": {
      "items": {
       "$ref": "#/components/schemas/Issue"
      },
      "type": "array",
      "title": "Issues"
     },
     "conflict": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Conflict"
       },
       {
        "type": "null"
       }
      ]
     },
     "default_action": {
      "type": "string",
      "enum": [
       "skip",
       "update",
       "new"
      ],
      "title": "Default Action"
     }
    },
    "type": "object",
    "required": [
     "index",
     "draft",
     "status",
     "issues",
     "default_action"
    ],
    "title": "PreviewRow"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}